# libs/domains/framework/backend/feature-agent-controller/spec/openapi.yaml
openapi: 3.1.0
info:
  title: Agent Controller API
  version: 1.0.0
  description: HTTP API for managing clients and proxied agent operations
  contact:
    name: ForePath
    email: hi@forepath.io
    url: https://forepath.io
servers:
  - url: http://localhost:3100/api
    description: Local development
security:
  - bearerAuth: []
paths:
...
  /clients/{id}/agents/{agentId}/vcs/status:
    get:
      summary: Get git status (proxied)
      operationId: getClientAgentGitStatus
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the client
        - in: path
          name: agentId
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the agent
      responses:
        '200':
          description: Git status information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GitStatusDto'
        '403':
          description: User does not have access to this client
        '404':
          description: Client or agent not found
  /clients/{id}/agents/{agentId}/vcs/branches:
    get:
      summary: List all branches (proxied)
      operationId: getClientAgentGitBranches
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the client
        - in: path
          name: agentId
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the agent
      responses:
        '200':
          description: Array of branch information
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GitBranchDto'
        '403':
          description: User does not have access to this client
        '404':
          description: Client or agent not found
    post:
      summary: Create a new branch (proxied)
      operationId: createClientAgentGitBranch
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the client
        - in: path
          name: agentId
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBranchDto'
      responses:
        '201':
          description: Branch created successfully
        '400':
          description: Invalid branch name or parameters
        '403':
          description: User does not have access to this client
        '404':
          description: Client or agent not found
  /clients/{id}/agents/{agentId}/vcs/branches/{branch}:
    delete:
      summary: Delete a branch (proxied)
      operationId: deleteClientAgentGitBranch
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the client
        - in: path
          name: agentId
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the agent
        - in: path
          name: branch
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Branch deleted successfully
        '400':
          description: Cannot delete current branch
        '403':
          description: User does not have access to this client
        '404':
          description: Client, agent, or branch not found
  /clients/{id}/agents/{agentId}/vcs/branches/{branch}/switch:
    post:
      summary: Switch to a branch (proxied)
      operationId: switchClientAgentGitBranch
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the client
        - in: path
          name: agentId
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the agent
        - in: path
          name: branch
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Branch switched successfully
        '400':
          description: Invalid branch or uncommitted changes
        '403':
          description: User does not have access to this client
        '404':
          description: Client, agent, or branch not found
  /clients/{id}/agents/{agentId}/vcs/diff:
    get:
      summary: Get file diff (proxied)
      operationId: getClientAgentGitFileDiff
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the client
        - in: path
          name: agentId
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the agent
        - in: query
          name: path
          required: true
          schema:
            type: string
          description: File path relative to repository root
      responses:
        '200':
          description: File diff information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GitDiffDto'
        '400':
          description: Invalid file path
        '403':
          description: User does not have access to this client
        '404':
          description: Client, agent, or file not found
  /clients/{id}/agents/{agentId}/vcs/stage:
    post:
      summary: Stage files (proxied)
      operationId: stageClientAgentFiles
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the client
        - in: path
          name: agentId
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StageFilesDto'
      responses:
        '204':
          description: Files staged successfully
        '400':
          description: Invalid file paths
        '403':
          description: User does not have access to this client
        '404':
          description: Client or agent not found
  /clients/{id}/agents/{agentId}/vcs/unstage:
    post:
      summary: Unstage files (proxied)
      operationId: unstageClientAgentFiles
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the client
        - in: path
          name: agentId
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnstageFilesDto'
      responses:
        '204':
          description: Files unstaged successfully
        '400':
          description: Invalid file paths
        '403':
          description: User does not have access to this client
        '404':
          description: Client or agent not found
  /clients/{id}/agents/{agentId}/vcs/commit:
    post:
      summary: Commit staged changes (proxied)
      operationId: commitClientAgentChanges
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the client
        - in: path
          name: agentId
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommitDto'
      responses:
        '204':
          description: Changes committed successfully
        '400':
          description: No staged changes or invalid commit message
        '403':
          description: User does not have access to this client
        '404':
          description: Client or agent not found
  /clients/{id}/agents/{agentId}/vcs/push:
    post:
      summary: Push changes to remote (proxied)
      operationId: pushClientAgentChanges
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the client
        - in: path
          name: agentId
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the agent
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PushOptionsDto'
      responses:
        '204':
          description: Changes pushed successfully
        '400':
          description: Nothing to push or push failed
        '403':
          description: User does not have access to this client
        '404':
          description: Client or agent not found
  /clients/{id}/agents/{agentId}/vcs/pull:
    post:
      summary: Pull changes from remote (proxied)
      operationId: pullClientAgentChanges
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the client
        - in: path
          name: agentId
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the agent
      responses:
        '204':
          description: Changes pulled successfully
        '400':
          description: Pull failed or conflicts
        '403':
          description: User does not have access to this client
        '404':
          description: Client or agent not found
  /clients/{id}/agents/{agentId}/vcs/fetch:
    post:
      summary: Fetch changes from remote (proxied)
      operationId: fetchClientAgentChanges
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the client
        - in: path
          name: agentId
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the agent
      responses:
        '204':
          description: Changes fetched successfully
        '403':
          description: User does not have access to this client
        '404':
          description: Client or agent not found
  /clients/{id}/agents/{agentId}/vcs/workspace/prepare-clean:
    post:
      summary: Prepare clean workspace (fetch, reset --hard, clean) — proxied
      operationId: prepareCleanClientAgentWorkspace
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: agentId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [baseBranch]
              properties:
                baseBranch:
                  type: string
      responses:
        '204':
          description: Workspace reset
        '403':
          description: Forbidden
        '404':
          description: Not found
...
  /clients/{id}/agents/{agentId}/vcs/rebase:
    post:
      summary: Rebase current branch (proxied)
      operationId: rebaseClientAgentBranch
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the client
        - in: path
          name: agentId
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RebaseDto'
      responses:
        '204':
          description: Rebase completed successfully
        '400':
          description: Rebase failed or conflicts
        '403':
          description: User does not have access to this client
        '404':
          description: Client, agent, or branch not found
  /clients/{id}/agents/{agentId}/vcs/conflicts/resolve:
    post:
      summary: Resolve merge conflict (proxied)
      operationId: resolveClientAgentConflict
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the client
        - in: path
          name: agentId
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResolveConflictDto'
      responses:
        '204':
          description: Conflict resolved successfully
        '400':
          description: Invalid conflict resolution strategy
        '403':
          description: User does not have access to this client
        '404':
          description: Client, agent, or file not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    GitStatusDto:
      type: object
      required:
        [
          currentBranch,
          isClean,
          hasUnpushedCommits,
          aheadCount,
          behindCount,
          files,
        ]
      properties:
        currentBranch:
          type: string
          description: Current branch name
        isClean:
          type: boolean
          description: Whether the working tree is clean (no changes)
        hasUnpushedCommits:
          type: boolean
          description: Whether there are unpushed commits (ahead of remote)
        aheadCount:
          type: integer
          description: Number of commits ahead of remote
        behindCount:
          type: integer
          description: Number of commits behind remote
        files:
          type: array
          items:
            $ref: '#/components/schemas/GitFileStatusDto'
          description: List of files with their git status
    GitFileStatusDto:
      type: object
      required: [path, status, type]
      properties:
        path:
          type: string
          description: File path relative to repository root
        status:
          type: string
          description: Git status code (e.g., 'M', 'A', 'D', '??', 'MM', etc.)
        type:
          type: string
          enum: [staged, unstaged, untracked, both]
          description: Status type
        isBinary:
          type: boolean
          description: Whether this is a binary file
        size:
          type: integer
          description: File size in bytes (for binary files)
    GitBranchDto:
      type: object
      required: [name, ref, isCurrent, isRemote, commit, message]
      properties:
        name:
          type: string
          description: Branch name (without refs/heads/ prefix)
        ref:
          type: string
          description: Full branch reference (e.g., 'refs/heads/main')
        isCurrent:
          type: boolean
          description: Whether this is the current branch
        isRemote:
          type: boolean
          description: Whether this is a remote branch
        remote:
          type: string
          description: Remote name (if remote branch)
        commit:
          type: string
          description: Commit hash (short)
        message:
          type: string
          description: Commit message (first line)
        aheadCount:
          type: integer
          description: Number of commits ahead of remote (for local branches)
        behindCount:
          type: integer
          description: Number of commits behind remote (for local branches)
    GitDiffDto:
      type: object
      required: [path, originalContent, modifiedContent, encoding, isBinary]
      properties:
        path:
          type: string
          description: File path relative to repository root
        originalContent:
          type: string
          description: Original file content (base64-encoded for text, empty for binary)
        modifiedContent:
          type: string
          description: Modified file content (base64-encoded for text, empty for binary)
        encoding:
          type: string
          enum: [utf-8, base64]
          description: Encoding indicator
        isBinary:
          type: boolean
          description: Whether this is a binary file
        originalSize:
          type: integer
          description: Original file size in bytes (for binary files)
        modifiedSize:
          type: integer
          description: Modified file size in bytes (for binary files)
    StageFilesDto:
      type: object
      required: [files]
      properties:
        files:
          type: array
          items:
            type: string
          description: Array of file paths to stage (empty array stages all)
    UnstageFilesDto:
      type: object
      required: [files]
      properties:
        files:
          type: array
          items:
            type: string
          description: Array of file paths to unstage (empty array unstages all)
    CommitDto:
      type: object
      required: [message]
      properties:
        message:
          type: string
          description: Commit message
    PushOptionsDto:
      type: object
      properties:
        force:
          type: boolean
          description: >
            When true, performs a `git push --force-with-lease`. Defaults to false.
    CreateBranchDto:
      type: object
      required: [name]
      properties:
        name:
          type: string
          description: Branch name (will be prefixed with conventional commit prefix if useConventionalPrefix is true)
        useConventionalPrefix:
          type: boolean
          description: Whether to use conventional commit prefix (feat/, fix/, chore/, etc.)
        conventionalType:
          type: string
          enum: [feat, fix, chore, docs, style, refactor, test, perf]
          description: Conventional commit type (only used if useConventionalPrefix is true)
        baseBranch:
          type: string
          description: Base branch to create from (defaults to current branch)
    RebaseDto:
      type: object
      required: [branch]
      properties:
        branch:
          type: string
          description: Branch name to rebase onto
    ResolveConflictDto:
      type: object
      required: [path, strategy]
      properties:
        path:
          type: string
          description: File path with conflict (relative to repository root)
        strategy:
          type: string
          enum: [yours, mine, both]
          description: Merge strategy - 'yours' (accept incoming), 'mine' (accept current), 'both' (keep both)
...